backend/requirements.txt
alembic==1.15.2
blinker==1.9.0
boto3==1.38.8
botocore==1.38.8
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
ecdsa==0.19.1
Flask==3.1.0
flask-cors==5.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
greenlet==3.2.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
Mako==1.3.10
MarkupSafe==3.0.2
pillow==11.2.1
psycopg2-binary==2.9.10
pyasn1==0.4.8
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
requests==2.32.3
rsa==4.9.1
s3transfer==0.12.0
six==1.17.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
